Remove redundant stuff, minor upgrades to work with EC2 and add github actions(Weekly CI)

+++ install.sh
@@ -1,24 +1,8 @@
-#####################################################
-#
-#          Personal Environment Config
-#
-#                     .--.
-#           ::\`--._,'.::.`._.--'/::
-#           ::::.  ` __::__ '  .::::
-#           ::::::-:.`'..`'.:-::::::
-#           ::::::::\ `--' /::::::::
-#                     '--'
-#
-#                  Try not.
-#         Do or do not. There is no try.
-#####################################################
-
-
 URL=https://github.com/ms-google/config.git
 CONFIGDIR=$HOME/.mconfig
 
 basic_update () {
-    sudo apt install -y --no-install-recommends git unzip wget
+    sudo apt install -y --no-install-recommends git unzip wget git-delta
     if [ ! -d "$CONFIGDIR" ] ; then
         echo "Config repo doesn't exist at $CONFIGDIR, cloning";
 
@@ -34,9 +18,6 @@ basic_update () {
             mv "$extracted_dir"/* "$dest_dir"
             rm -r "$extracted_dir"
         fi
-
-
-        # git clone "$URL" "$CONFIGDIR";
     else
         echo "Config repo exists at $CONFIGDIR, checking for updates";
         cd "$CONFIGDIR";
@@ -45,41 +26,44 @@ basic_update () {
 }
 
 dependencies_ubuntu () {
-    # Install neovim / vim 8.0
-    sudo apt install -y --no-install-recommends neovim vim
-    sudo apt install -y --no-install-recommends ncdu tmux ranger w3m curl htop
+    sudo add-apt-repository -y ppa:neovim-ppa/stable
+    sudo apt update -y
+    sudo apt install -y --no-install-recommends neovim vim ranger tmux
+    sudo apt install -y --no-install-recommends ncdu wget w3m curl htop btop ripgrep bat nvtop
     sudo apt install -y --no-install-recommends xclip xsel
+
+    if which python &>/dev/null; then
+        python -m ensurepip --upgrade
+        python -m pip install uv
+        uv pip install --system beautifulsoup4 black flake8 ipython isort matplotlib-inline numpy pandas requests rich ruff uploadserver uv yt-dlp
+    else
+        echo "Python not found, skipping command."
+    fi
+
     echo "Dependencies installed";
 }
 
 bin_update() {
     sudo cp $CONFIGDIR/git/redate /usr/bin/redate && sudo chmod +x /usr/bin/redate
-    sudo cp $CONFIGDIR/python/loadpy /usr/bin/loadpy && sudo chmod +x /usr/bin/loadpy
     echo "Binaries installed";
 }
 
-config_update() {
-    cp $CONFIGDIR/git/gitconfig $HOME/.gitconfig
-    mkdir -p $HOME/.ipython/ && cp $CONFIGDIR/python/ipython_config $HOME/.ipython/ipython_config.py
-    mkdir -p $HOME/.config/ranger && cp $CONFIGDIR/ranger/* $HOME/.config/ranger/
-    echo "configs updated";
-}
-
 zsh_update() {
-    sudo apt install -y zsh fonts-font-awesome fzf
+    sudo apt install -y --no-install-recommends zsh fonts-font-awesome fzf
     sudo chsh -s $(which zsh)
 
     git clone https://github.com/ohmyzsh/ohmyzsh.git ~/.oh-my-zsh
-    git clone --depth=1 https://github.com/romkatv/powerlevel10k.git ~/powerlevel10k
+    git clone --depth=1 https://github.com/romkatv/powerlevel10k.git ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/themes/powerlevel10k
     git clone https://github.com/zsh-users/zsh-syntax-highlighting.git $HOME/.oh-my-zsh/custom/plugins/zsh-syntax-highlighting
     git clone https://github.com/zsh-users/zsh-autosuggestions.git $HOME/.oh-my-zsh/custom/plugins/zsh-autosuggestions
 
-    mv $HOME/.zshrc $HOME/.zshrc_old
+    if test -f $HOME/.zshrc; then
+        mv $HOME/.zshrc $HOME/.zshrc_old
+    fi
     cp $CONFIGDIR/zsh/zshrc $HOME/.zshrc && cp $CONFIGDIR/zsh/aliases $HOME/.zsh_aliases
-    cp $CONFIGDIR/zsh/p10k.zsh $HOME/.p10k.zsh
 
-    sudo mkdir -p /usr/share/fonts/truetype/fonts-iosveka/
-    sudo cp $CONFIGDIR/static/iosveka-regular.ttf /usr/share/fonts/truetype/fonts-iosveka/
+    sudo mkdir -p /usr/share/fonts/truetype/fonts-iosevka/
+    sudo cp $CONFIGDIR/static/iosevka-regular.ttf /usr/share/fonts/truetype/fonts-iosevka/
     echo "zsh updated";
 }
 
@@ -87,7 +71,7 @@ tmux_update() {
     mkdir -p $HOME/.tmux/
     cp tmux/tmux.conf $HOME/.tmux.conf
     git clone https://github.com/tmux-plugins/tpm $HOME/.tmux/plugins/tpm
-    tmux source-file $HOME/.tmux.conf
+    tmux new-session -d && tmux source-file $HOME/.tmux.conf
     echo "tmux updated";
 }
 
@@ -106,6 +90,7 @@ i3_update() {
     sudo apt-get install -y --no-install-recommends py3status i3 i3lock rofi
     mkdir -p $HOME/.config/i3
     cp $CONFIGDIR/i3/*confg $HOME/.config/i3/
+    cp $CONFIGDIR/i3/*.py /usr/bin/
     echo "i3 updated";
 }
 
@@ -124,7 +109,6 @@ vim_update() {
 common_update() {
     basic_update
     dependencies_ubuntu
-    config_update
     bin_update
     zsh_update
     vim_update
